+++ src/main/kotlin/com/devfloor/untitled/article/application/ArticleService.kt
@@ -2,7 +2,7 @@ package com.devfloor.untitled.article.application
 
 import com.devfloor.untitled.article.domain.Article
 import com.devfloor.untitled.article.domain.ArticleRepository
-import com.devfloor.untitled.exception.domain.NotFoundException
+import com.devfloor.untitled.common.exception.NotFoundException
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
