+++ src/main/kotlin/com/devfloor/untitled/common/exception/NotFoundException.kt
@@ -1,4 +1,4 @@
-package com.devfloor.untitled.exception.domain
+package com.devfloor.untitled.common.exception
 
 data class NotFoundException(
     override val message: String?
