+++ src/main/kotlin/com/devfloor/untitled/common/presentation/AdviceController.kt
@@ -1,13 +1,13 @@
-package com.devfloor.untitled.exception.presentation
+package com.devfloor.untitled.common.presentation
 
-import com.devfloor.untitled.exception.domain.NotFoundException
+import com.devfloor.untitled.common.exception.NotFoundException
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.ControllerAdvice
 import org.springframework.web.bind.annotation.ExceptionHandler
 
 @ControllerAdvice
-class ExceptionController {
+class AdviceController {
     @ExceptionHandler(value = [NotFoundException::class])
     fun handleNoSuchElementException(): ResponseEntity<Any> =
         ResponseEntity(HttpStatus.NOT_FOUND)
